ScanCellObjects
Add two new functions to ScanCellObjects that work fine as scripts but are too slow and have to be converted to native.

Add another condition to the native ScanCellObjects function to check for the BaseForm property since developing the new functions I noticed that without that use to return a loot of incorrect objects.